[ADD] RemoveWorkspaceUseCase

+++ Notey/Data/Storage/Workspace/Local/LocalWorkspaceStorage.swift
@@ -19,13 +19,16 @@ public protocol LocalWorkspaceStorage: WorkspaceStorage {
     @discardableResult
     func removeAllWorkspace() -> Observable<[WorkspaceDomain]>
     
+    @discardableResult
+    func removeWorkspace(_ object: WorkspaceDomain) -> Observable<WorkspaceDomain>
+    
 }
 
 public final class DefaultLocalWorkspaceStorage {
     
     let coreDataStorage: CoreDataStorageShared
     
-    init(coreDataStorage: CoreDataStorageShared = CoreDataStorage.shared) {
+    public init(coreDataStorage: CoreDataStorageShared = CoreDataStorage.shared) {
         self.coreDataStorage = coreDataStorage
     }
     
@@ -34,7 +37,7 @@ public final class DefaultLocalWorkspaceStorage {
 extension DefaultLocalWorkspaceStorage: LocalWorkspaceStorage {
     
     public func fetchAllWorkspace() -> Observable<[WorkspaceDomain]> {
-        return Observable<[WorkspaceDomain]>.create { [unowned self] (observer) -> Disposable in
+        return Observable<[WorkspaceDomain]>.create { (observer) -> Disposable in
             self.coreDataStorage.performBackgroundTask { (context) in
                 do {
                     let request: NSFetchRequest = WorkspaceEntity.fetchRequest()
@@ -51,7 +54,7 @@ extension DefaultLocalWorkspaceStorage: LocalWorkspaceStorage {
                 }
             }
             return Disposables.create()
-        }.observeOn(CurrentThreadScheduler.instance)
+        }.observeOn(MainScheduler.instance)
     }
     
     public func insertSynchronizeWorkspace(_ object: WorkspaceDomain) -> Observable<WorkspaceDomain> {
@@ -62,7 +65,7 @@ extension DefaultLocalWorkspaceStorage: LocalWorkspaceStorage {
                     let exisitingEntities = try context.fetch(request)
                     let inserted: WorkspaceEntity
                     if let coreId = object.coreId,
-                       let existing = exisitingEntities.first(where: { $0.objectID.uriRepresentation().path == coreId }) {
+                       let existing = exisitingEntities.first(where: { $0.objectID == coreId }) {
                         inserted = existing.synchronizeWithObjectId(with: object, context: context)
                     } else {
                         inserted = WorkspaceEntity(object, insertInto: context)
@@ -86,13 +89,46 @@ extension DefaultLocalWorkspaceStorage: LocalWorkspaceStorage {
             self.coreDataStorage.performBackgroundTask { (context) in
                 do {
                     let request: NSFetchRequest = WorkspaceEntity.fetchRequest()
-                    let workspaces = try context.fetch(request)
-                    workspaces.forEach { (element) in
-                        context.delete(element)
-                    }
+                    let entities = try context.fetch(request)
+                    entities.forEach { context.delete($0) }
                     try context.save()
-                    let domains = workspaces.map { $0.toDomain() }
+                    let domains = entities.map { $0.toDomain() }
                     observer.onNext(domains)
+                    observer.onCompleted()
+                } catch {
+                    let error = CoreDataStorageError.deleteError(error)
+                    observer.onError(error)
+                    observer.onCompleted()
+                }
+            }
+            return Disposables.create()
+        }.observeOn(CurrentThreadScheduler.instance)
+    }
+    
+    public func removeWorkspace(_ object: WorkspaceDomain) -> Observable<WorkspaceDomain> {
+        return Observable<WorkspaceDomain>.create { [unowned self] (observer) -> Disposable in
+            self.coreDataStorage.performBackgroundTask { (context) in
+                guard let coreId = object.coreId else {
+                    let message = "Failed to execute removeWorkspace(_:) caused by coreId is not available"
+                    let error = NSError(domain: message, code: 0, userInfo: nil)
+                    observer.onError(error)
+                    observer.onCompleted()
+                    return
+                }
+                
+                let removedObject = context.object(with: coreId)
+                guard let entity = removedObject as? WorkspaceEntity else {
+                    let message = "Failed to execute removeWorkspace(_:) caused by removedObject (\(removedObject)) is not WorkspaceEntity"
+                    let error = NSError(domain: message, code: 0, userInfo: nil)
+                    observer.onError(error)
+                    observer.onCompleted()
+                    return
+                }
+                
+                do {
+                    context.delete(entity)
+                    try context.save()
+                    observer.onNext(object)
                     observer.onCompleted()
                 } catch {
                     let error = CoreDataStorageError.deleteError(error)
